Audio fix for qsd8k devices
Update to libaudio and libaudiopolicy for compatibility with ICS.
Name space has been changed from android to android_audio_legacy.
Module names have been updated to new ICS standards.
Cleanup makefile.

Credit to sean.neeley@gmail.com for correcting openOutputStream()
It seems channels & sampleRate change each time the code is called
causing audioflinger to error. For instance it asked for a stereo
audio stream and got back a mono one because we created a mono one
first and gave it the same mono one the second time the code was called.
so before returning the same old output stream if it already exists.
First call set() on it with the new values for channels and sampleRate.

Change-Id: If7f68b846ef8fbd8aec817a6da10e49c10c7a732